requirements.txt
requests==2.28.2
pandas==2.2.2
mysql.connector==2.2.9
charset-normalizer==3.1.0